backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
qdrant-client
markitdown==0.0.1a2
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
httpx==0.25.2

# Database migrations
alembic==1.12.1

# Additional Pydantic features
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Authentication and security
passlib[bcrypt]
python-jose[cryptography]==3.3.0

# File handling
aiofiles==23.2.0
python-magic==0.4.27

# Testing (development)
pytest==7.4.0
pytest-asyncio==0.21.1
aiohttp==3.9.1
pytest-cov==4.1.0

# Vector embeddings and text processing
numpy>=1.24.0
tiktoken>=0.5.0